move everything inside src folder except main.go and yaml
