+++ src/slideshow/slideshow_test.go
